auth_service/requirements.txt
boto3
fastapi
pyjwt
uvicorn
pydantic[email]
